// web/interfaces/store.ts
export interface StoreDetail {
  ID: number;
  CompanyID: number;
  Name: string;
  IsActive: boolean;
  OpenStatus: string;
  BusinessTypeCode: string;
  ContractPlanCode: string;
  DistrictID: number;
  DistrictName: string;
  PrefectureID: number;
  PrefectureName: string;
  RegionID: number;
  RegionName: string;
}